Feature/generic GitHub action

+++ src/main/java/net/lego/data/v1/dto/Condition.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 @Builder
 public class Condition {
-    private Long conditionId;
+    private Integer conditionId;
     private String conditionCode;
     private String conditionDescription;
     private String conditionText;
